gth-browser.c: Improve handling of invalid uris.
Normally when one attempts to start pix with a bad uri, you only
get an info bar and a blank view. This ends up breaking all
navigation buttons as well, making it difficult state to recover
from. This is made worse because pix also remembers the last
location and tries to display it at startup.

Instead, if a location fails to load, show the home folder, but still
show the info bar reporting the problem.